examples/basic/frameworks/background_agent_v2/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools >= 64", "setuptools-scm>=8"]

[tool.setuptools_scm]
root = "../../../.."

[project]
name = "aiq_multi_frameworks"
dynamic = ["version"]
dependencies = [
  "aiqtoolkit[langchain,llama-index]~=1.2",
  "arxiv~=2.1.3",
  "bs4==0.0.2",
  "GitPython~=3.1.0",
  "markdown-it-py~=3.0",
  "nvidia-haystack==0.1.2",
  "openai~=1.78.0",  # Pin to compatible range with llama-index
  "requests~=2.32.0",
]
requires-python = ">=3.11,<3.13"
description = "Custom AIQ toolkit Workflow"
classifiers = ["Programming Language :: Python"]

[tool.uv.sources]
aiqtoolkit = { path = "../../../..", editable = true }

[project.entry-points.'aiq.components']
aiq_multi_frameworks = "aiq_multi_frameworks.register"
